Moved some files into new folders and changed some parameters
