+++ app/src/main/java/com/example/rebootapp/Adapters/GameListAdapter.java
@@ -1,4 +1,4 @@
-package com.example.rebootapp.ListFiles;
+package com.example.rebootapp.Adapters;
 
 import android.app.Instrumentation;
 import android.content.Context;
