+++ app/src/main/java/com/example/rebootapp/Adapters/GameModel.java
@@ -1,4 +1,4 @@
-package com.example.rebootapp.ListFiles;
+package com.example.rebootapp.Adapters;
 
 
 import android.util.Log;
