+++ app/src/main/java/com/example/rebootapp/Adapters/GameSearchAdapter.java
@@ -1,4 +1,4 @@
-package com.example.rebootapp;
+package com.example.rebootapp.Adapters;
 
 import android.content.Context;
 import android.content.Intent;
@@ -14,6 +14,9 @@ import androidx.annotation.NonNull;
 import androidx.recyclerview.widget.RecyclerView;
 
 import com.bumptech.glide.Glide;
+import com.example.rebootapp.GameSearch;
+import com.example.rebootapp.GameSearchDetailsActivity;
+import com.example.rebootapp.R;
 
 import org.parceler.Parcels;
