+++ app/src/main/java/com/example/rebootapp/Adapters/ManageListAdapter.java
@@ -1,4 +1,4 @@
-package com.example.rebootapp.ListFiles;
+package com.example.rebootapp.Adapters;
 
 import android.content.Context;
 import android.content.DialogInterface;
