+++ app/src/main/java/com/example/rebootapp/Adapters/UserListModel.java
@@ -1,4 +1,4 @@
-package com.example.rebootapp.ListFiles;
+package com.example.rebootapp.Adapters;
 
 import java.util.List;
