+++ app/src/main/java/com/example/rebootapp/Fragments/SearchFragment.java
@@ -18,7 +18,7 @@ import androidx.recyclerview.widget.RecyclerView;
 import com.codepath.asynchttpclient.AsyncHttpClient;
 import com.codepath.asynchttpclient.callback.JsonHttpResponseHandler;
 import com.example.rebootapp.GameSearch;
-import com.example.rebootapp.GameSearchAdapter;
+import com.example.rebootapp.Adapters.GameSearchAdapter;
 import com.example.rebootapp.R;
 
 import org.json.JSONArray;
