+++ app/src/main/java/com/example/rebootapp/FriendsActivity.java
@@ -1,39 +1,25 @@
 package com.example.rebootapp;
 
-import static androidx.test.InstrumentationRegistry.getContext;
-
 import androidx.appcompat.app.AppCompatActivity;
 
-import android.content.Intent;
 import android.os.Bundle;
 import android.util.Log;
 import android.view.View;
-import android.widget.ImageView;
-import android.widget.RatingBar;
-import android.widget.TextView;
 import android.widget.Button;
 import android.widget.EditText;
 import android.widget.ImageButton;
-import android.widget.ImageView;
 import android.widget.Toast;
-import android.widget.ToggleButton;
 
-import androidx.recyclerview.widget.GridLayoutManager;
 import androidx.recyclerview.widget.LinearLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
 
 
-import com.bumptech.glide.Glide;
-import com.parse.CountCallback;
+import com.example.rebootapp.Adapters.FriendsListAdapter;
 import com.parse.FindCallback;
-import com.parse.GetCallback;
-import com.parse.ParseObject;
 import com.parse.ParseException;
 import com.parse.ParseFile;
 import com.parse.ParseQuery;
 import com.parse.ParseUser;
-
-import org.parceler.Parcels;
 
 import java.util.ArrayList;
 import java.util.List;
